Reimplement in plain C
The code is now much smaller when we get rid of the arduino stuff, it
might possibly even fit on an attiny13 now. In addition the timing is
much more stable now.

A makefile was added, it supports programming the device using a buspirate
board. The CPU is set to 8 MHz, which already gives a better accuracy than
the arduino code since the measurements are not polluted by interrupts, and
it remains compatible with small boards like the trinket.