Add support for a 2nd function : front lights
Some remotes have a 3rd channel supporting lights or whatever. This
channel usually presents either a short or a large pulse that a
decode can use. Here we simply enable PB0 (pin 5) if the pulse is
larger than 1400 us. It allows this device to be used for this sole
purpose.