Add support for rear and brake leds
These ones are turned on during BRK and REV. They're connected to the
same pin (PB2, pin 7). The 2 red leds are connected in series to GND,
the white led to VCC. Their series voltage is not enough to turn all
of them on, so it's possible to turn either one or the other individually.