change GPIO assignment to ease soldering
The previous GPIO assignment was not convenient for soldering because
the IN and OUT connectors had to be placed on the same side.

It's important to keep in mind that PB3 can't be used as input on the
trinket, apparently due to the USB resistors and zeners. Thus we return
the signals so that the input comes from PB2 (pin7) and the output goes
through PB3(pin2). PB1(pin6) is not changed so that it continues to
match the debug led on the trinket.

Now the IN and OUT are inline, the rear lights are on the same side as
the output, and the front light is on the input side :

    PB#  pin  dir   role
   -----+---+-----+--------------
    PB0 | 5 | out | (front light)
    PB1 | 6 | out | (debug)
    PB2 | 7 |  in | (pulse-in)
    PB3 | 2 | out | (pulse-out)
    PB4 | 3 | i/o | (brake/rear)
   -----+---+-----+--------------